Doctor Kafka enhancement: adding the capabilty to fetch offset/data
Currently Dr. Kafka only checks port 9092 to verify whether a broker is live or dead.  But in opeartion, we saw quite a few times that a broker is hanging but can still show port 9092 as active.

Added more checks (fetch an offset and then fetch data from the offset for that offset) to verfy whether the broker is really dead or alive.

On the next checkin, I will add some automatic active code for dead broker